requirement.txt
anyio==3.7.1
argcomplete==3.6.2
argh==0.31.3
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///croot/argon2-cffi-bindings_1736182440035/work
arrow==1.3.0
asttokens @ file:///croot/asttokens_1743630435401/work
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach @ file:///croot/bleach_1732290411627/work
bokeh==3.7.3
certifi==2025.4.26
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
debugpy==1.8.14
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
einops==0.8.1
entrypoints @ file:///work/ci_py311/entrypoints_1676823319002/work
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
flash_attn==2.8.0.post2
fonttools==4.58.1
fqdn==1.5.1
fsspec==2025.5.1
gffutils==0.13
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
intervaltree==3.1.0
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1744912971218/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipython_pygments_lexers @ file:///croot/ipython_pygments_lexers_1744753235686/work
isoduration==20.11.0
jedi @ file:///croot/jedi_1733987392413/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
json5==0.12.0
jsonpointer==3.0.0
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter-events @ file:///croot/jupyter_events_1741184577592/work
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.8.1
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.16.0
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1744706698694/work
jupyterlab==4.4.3
jupyterlab_pygments @ file:///croot/jupyterlab_pygments_1741124142640/work
jupyterlab_server==2.27.3
jupyterthemes==0.20.0
kiwisolver==1.4.8
lesscpy==0.15.1
livelossplot==0.5.6
llvmlite==0.44.0
lxml==5.4.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.3
matplotlib-inline @ file:///work/ci_py311/matplotlib-inline_1676823841154/work
mistune @ file:///croot/mistune_1741124011532/work
mpmath==1.3.0
narwhals==1.43.1
nb-conda==2.2.1
nb_conda_kernels @ file:///croot/nb_conda_kernels_1744229677905/work
nbclassic @ file:///croot/nbclassic_1717622582894/work
nbclient @ file:///croot/nbclient_1741123995822/work
nbconvert @ file:///croot/nbconvert-meta_1741184653315/work
nbformat @ file:///croot/nbformat_1728049424075/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
networkx==3.4.2
ninja==1.11.1.4
notebook==7.4.3
notebook_shim @ file:///croot/notebook-shim_1741707758683/work
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pillow==11.2.1
platformdirs @ file:///croot/platformdirs_1744273042065/work
ply==3.11
prometheus_client @ file:///croot/prometheus_client_1744271615306/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyahocorasick==2.2.0
pycparser==2.22
pyfaidx==0.8.1.4
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.2.3
pysam==0.21.0
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger==3.3.0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1736541261634/work
seaborn==0.13.2
Send2Trash==1.8.3
simplejson==3.20.1
six @ file:///croot/six_1744271502820/work
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
tinycss2 @ file:///croot/tinycss2_1738337643607/work
torch==2.6.0
tornado @ file:///croot/tornado_1748956929273/work
tqdm==4.67.1
traitlets @ file:///croot/traitlets_1718227057033/work
triton==3.2.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==24.11.1
webencodings==0.5.1
websocket-client @ file:///croot/websocket-client_1715878298792/work
xyzservices==2025.4.0
